refractor: minor changes

+++ lib/models/world_stats.dart
@@ -9,7 +9,7 @@ class WorldStatsModel {
   int? active;
   int? critical;
   int? casesPerOneMillion;
-  int? deathsPerOneMillion;
+  double? deathsPerOneMillion;
   int? tests;
   double? testsPerOneMillion;
   int? population;
